refactor zynqmon task code
